+++ CHANGELOG.md
@@ -6,6 +6,8 @@ a Changelog](https://keepachangelog.com/en/1.0.0/) and the project adheres to
 [semver](https://semver.org/).
 
 ## [UNRELEASED]
+
+## [0.3.0] - 2020-11-13
 Added:
 - Error handling for every error thrown in an action, not just `failure(...)`
   return values.
@@ -38,6 +40,7 @@ Fixed:
 ## [0.1.0] - 2020-02-19
 Initial release.
 
-[UNRELEASED]: https://github.com/PsychoLlama/retreon/compare/v0.2.0...HEAD
+[UNRELEASED]: https://github.com/PsychoLlama/retreon/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/PsychoLlama/retreon/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/PsychoLlama/retreon/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/PsychoLlama/retreon/releases/tag/v0.1.0